+++ NEWS.md
@@ -23,6 +23,7 @@ Enhancements:
   * Prevent unhandled crashes if the 'replicates' argument is badly structured.
   * Show number of released fish per release location in the release sites table.
   * Detailed progression per group and release site is now exported through a `release.overview` object.
+  * Prevent migration from continuing if some arrays are not associated to a section.
 
 
 ## actel 0.1.2